+++ subscriptions/client/database/subscriptions.go
@@ -2,9 +2,9 @@ package database
 
 import "main/models"
 
+// GetSubscriptions gets a list of all subscriptions from the database
 func (c *Client) GetSubscriptions() (subscriptions []models.UserSubscription, err error) {
-
-	rows, err := c.DB.Query("SELECT subscription_id, status, charge_amount FROM subscriptions")
+	rows, err := c.DB.Query("SELECT user_id, subscription_id, status, charge_amount FROM subscriptions")
 	if err != nil {
 		return nil, err
 	}
@@ -12,7 +12,7 @@ func (c *Client) GetSubscriptions() (subscriptions []models.UserSubscription, er
 
 	for rows.Next() {
 		subscription := models.UserSubscription{}
-		if err := rows.Scan(&subscription.SubscriptionID, &subscription.Status, &subscription.ChargeAmount); err != nil {
+		if err := rows.Scan(&subscription.UserID, &subscription.SubscriptionID, &subscription.Status, &subscription.ChargeAmount); err != nil {
 			return nil, err
 		}
 		subscriptions = append(subscriptions, subscription)
@@ -20,12 +20,22 @@ func (c *Client) GetSubscriptions() (subscriptions []models.UserSubscription, er
 	return subscriptions, nil
 }
 
-func (c *Client) Subscribe(s models.UserSubscription) (subscriptionID string, err error) {
-	sqlStatementINSERT := "INSERT INTO subscriptions(user_id, subscription_id, status, charge_amount) VALUES ($1,$2,$3,$4) RETURNING subscription_id"
+// GetUserSubscriptions gets a list of a user's subscriptions from the database
+func (c *Client) GetUserSubscriptions(userID string) (subscriptions []models.UserSubscription, err error) {
+	sqlStatement := "SELECT user_id, subscription_id, status, charge_amount FROM subscriptions WHERE user_id = $1"
 
-	err = c.DB.QueryRow(sqlStatementINSERT, s.UserID, s.SubscriptionID, s.Status, s.ChargeAmount).Scan(&s.SubscriptionID)
+	rows, err := c.DB.Query(sqlStatement, userID)
 	if err != nil {
-		return subscriptionID, err
+		return nil, err
 	}
-	return subscriptionID, nil
+	defer rows.Close()
+
+	for rows.Next() {
+		subscription := models.UserSubscription{}
+		if err := rows.Scan(&subscription.UserID, &subscription.SubscriptionID, &subscription.Status, &subscription.ChargeAmount); err != nil {
+			return nil, err
+		}
+		subscriptions = append(subscriptions, subscription)
+	}
+	return subscriptions, err
 }
